refactor: :recycle: seperating example module from mediatr package
