+++ internal/examples/cqrs_example/docs/swagger.json
