+++ internal/examples/cqrs_example/internal/products/api/products_routes.go
