+++ internal/examples/cqrs_example/internal/products/dtos/product_dto.go
