+++ internal/examples/cqrs_example/internal/products/features/creating_product/commands/create_product.go
