+++ internal/examples/cqrs_example/internal/products/features/creating_product/dtos/create_product_request_dto.go
