+++ internal/examples/cqrs_example/internal/products/features/creating_product/events/product_created_handler.go
