+++ internal/examples/cqrs_example/internal/products/features/getting_product_by_id/dtos/get_product_by_id_request_dto.go
