+++ internal/examples/cqrs_example/internal/products/features/getting_product_by_id/dtos/get_product_by_id_response.go
@@ -1,8 +1,6 @@
 package dtos
 
-import (
-	"github.com/mehdihadeli/go-mediatr/examples/cqrs/internal/products/dtos"
-)
+import "cqrsexample/internal/products/dtos"
 
 type GetProductByIdQueryResponse struct {
 	Product *dtos.ProductDto `json:"product"`
