+++ internal/examples/cqrs_example/internal/products/features/getting_product_by_id/queries/get_product_by_id.go
