+++ internal/examples/cqrs_example/internal/products/models/product.go
@@ -1,8 +1,9 @@
 package models
 
 import (
-	uuid "github.com/satori/go.uuid"
 	"time"
+
+	uuid "github.com/satori/go.uuid"
 )
 
 // Product model
